factor out AdvOptSidebar
